+++ tests/test_statistics.py
